% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{new_explorer_board}
\alias{new_explorer_board}
\title{Create a Share Explorer Board}
\usage{
new_explorer_board(tickers = c("AAPL", "MSFT", "GOOG"), ...)
}
\arguments{
\item{tickers}{Character vector of initial tickers}

\item{...}{Forwarded to the dock board constructor}
}
\value{
A dock board
}
\description{
Two-block pipeline: ticker_data -> explorer
}
